HiKey.dsc: disable UEFI boot mode
For OP-TEE to work properly, the memory map has to be modified. I am doing it
via the device tree for now because I don't know how to do it in EDK2.
So I need to revert to using the memory map defined in the FDT and thus
disable booting with the UEFI stub (which strips memory info from the FDT and
populates EFI nodes instead, so that the kernel will query UEFI services to
learn the memory map).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>